fix: Better error message on binary failure
Show the user a better error message when the clang-format execSync
fails. Previoustly we were ignoring stderr because it didn't work on
windows, ive changed that and it seems to be working on my windows
machine so node must have fixed it? Also use path.join on the binary
file path to ensure consistent slashes.